Add components to routes
