+++ client/src/pages/CreateFilm.jsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export const CreateFilm = () => (
+    <>
+        Create film page
+    </>
+)
