+++ client/src/pages/FilmDetail.jsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export const FilmDetail = () => (
+    <>
+        Film detail page
+    </>
+)
