+++ client/src/routes.js
@@ -1,16 +1,19 @@
 import React from 'react'
 import { Switch, Route, Redirect } from 'react-router-dom'
+import { Films } from './pages/Films'
+import { FilmDetail } from './pages/FilmDetail'
+import { CreateFilm } from './pages/CreateFilm'
 
 export const useRoutes = () => (
     <Switch>
         <Route exact path="/">
-
-        </Route>
-        <Route path="/film/:id">
-
+            <Films />
         </Route>
         <Route exact path="/film/new">
-
+            <CreateFilm />
+        </Route>
+        <Route path="/film/:id">
+            <FilmDetail />
         </Route>
         <Redirect to="/" />
     </Switch>
